댓글 생성,삭제 구현

+++ cheerup_board/views.py
@@ -71,7 +71,6 @@ class update_board(UpdateView, ListView):
 
 def delete_board(request, pk): # have to make the password confirmation
     post = get_object_or_404(PhotoPost, pk=pk)
-    print(request.POST.get('anony_password'))
     if check_password(request.POST.get('anony_password'), post.anony_password):
         post.delete()
         return JsonResponse({'message': True })
@@ -101,11 +100,13 @@ class board_detail(DetailView):
 
     def post(self, request, *args, **kwargs):
         content = request.POST.get('content')
+        photopost = self.get_object()
+        anony_password = request.POST.get('anony_password')
         if content:
-            photopost = self.get_object()
-            comment = Comment(post=photopost, content=content)
+            anony_password_hashed = make_password(anony_password)
+            comment = Comment(post=photopost, anony_password=anony_password_hashed, content=content)
             comment.save()
-        return redirect(reverse('board:board_list'))
+        return redirect('board:board_detail', pk=photopost.pk)
 
 
 class update_comment(UpdateView):
@@ -119,8 +120,10 @@ class update_comment(UpdateView):
 
 def delete_comment(request, pk): # have to make the password confirmation
     comment = get_object_or_404(Comment, pk=pk)
-    comment.delete()
-    return redirect(reverse('board:board_list'))
+    if check_password(request.POST.get('anony_password'), comment.anony_password):
+        comment.delete()
+        return JsonResponse({'message': True })
+    return JsonResponse({'message': False })
 
 
 # don't use list view on comment, it appeared in PhotoPost detail view
@@ -155,8 +158,9 @@ class MessageView(View):
         author = get_api("https://nickname.hwanmoo.kr/?format=json&count=1")["words"][0]
         anony_password = request.POST.get('anony_password')
         content = request.POST.get('content')
-        anony_password_hashed = make_password(anony_password)
-        Message.objects.create(author=author, anony_password=anony_password_hashed, content=content)
+        if content:
+            anony_password_hashed = make_password(anony_password)
+            Message.objects.create(author=author, anony_password=anony_password_hashed, content=content)
         
         return redirect('board:message_list')
